+++ tilequeue/command.py
@@ -1,6 +1,6 @@
+from collections import defaultdict
 from collections import Iterable
 from collections import namedtuple
-from collections import defaultdict
 from contextlib import closing
 from itertools import chain
 from jinja2 import Environment
@@ -12,15 +12,16 @@ from tilequeue.format import lookup_format_by_extension
 from tilequeue.metro_extract import city_bounds
 from tilequeue.metro_extract import parse_metro_extract
 from tilequeue.query import DataFetcher
+from tilequeue.query import jinja_filter_bbox
 from tilequeue.query import jinja_filter_bbox_filter
 from tilequeue.query import jinja_filter_bbox_intersection
-from tilequeue.query import jinja_filter_bbox_padded_intersection
-from tilequeue.query import jinja_filter_bbox
-from tilequeue.query import jinja_filter_geometry
 from tilequeue.query import jinja_filter_bbox_overlaps
+from tilequeue.query import jinja_filter_bbox_padded_intersection
+from tilequeue.query import jinja_filter_geometry
 from tilequeue.queue import make_sqs_queue
 from tilequeue.store import s3_tile_key
 from tilequeue.tile import coord_int_zoom_up
+from tilequeue.tile import coord_is_valid
 from tilequeue.tile import coord_marshall_int
 from tilequeue.tile import coord_unmarshall_int
 from tilequeue.tile import create_coord
@@ -45,20 +46,20 @@ from tilequeue.postgresql import DBAffinityConnectionsNoLimit
 from urllib2 import urlopen
 from zope.dottedname.resolve import resolve
 import argparse
+import datetime
 import logging
 import logging.config
 import multiprocessing
 import operator
 import os
+import os.path
 import Queue
 import signal
 import sys
 import threading
 import time
-import yaml
-import datetime
-import os.path
 import traceback
+import yaml
 
 
 def create_command_parser(fn):
@@ -352,21 +353,25 @@ def tilequeue_intersect(cfg, peripherals):
 
     assert cfg.intersect_expired_tiles_location, \
         'Missing tiles expired-location configuration'
-    assert os.path.isdir(cfg.intersect_expired_tiles_location), \
-        'tiles expired-location is not a directory'
+    assert os.path.exists(cfg.intersect_expired_tiles_location), \
+        'tiles expired-location does not exist'
 
-    file_names = os.listdir(cfg.intersect_expired_tiles_location)
-    if not file_names:
-        logger.info('No expired tiles found, terminating.')
-        return
-    file_names.sort()
-    # cap the total number of files that we process in one shot
-    # this will limit memory usage, as well as keep progress moving
-    # along more consistently rather than bursts
-    expired_tile_files_cap = 20
-    file_names = file_names[:expired_tile_files_cap]
-    expired_tile_paths = [os.path.join(cfg.intersect_expired_tiles_location, x)
-                          for x in file_names]
+    if os.path.isdir(cfg.intersect_expired_tiles_location):
+        file_names = os.listdir(cfg.intersect_expired_tiles_location)
+        if not file_names:
+            logger.info('No expired tiles found, terminating.')
+            return
+        file_names.sort()
+        # cap the total number of files that we process in one shot
+        # this will limit memory usage, as well as keep progress moving
+        # along more consistently rather than bursts
+        expired_tile_files_cap = 20
+        file_names = file_names[:expired_tile_files_cap]
+        expired_tile_paths = \
+            [os.path.join(cfg.intersect_expired_tiles_location, x)
+             for x in file_names]
+    else:
+        expired_tile_paths = [cfg.intersect_expired_tiles_location]
 
     logger.info('Fetching tiles of interest ...')
     tiles_of_interest = peripherals.toi.fetch_tiles_of_interest()
@@ -422,7 +427,7 @@ def tilequeue_intersect(cfg, peripherals):
 def make_store(store_type, store_name, cfg):
     if store_type == 'directory':
         from tilequeue.store import make_tile_file_store
-        return make_tile_file_store(store_name)
+        return make_tile_file_store(cfg.s3_path or store_name)
 
     elif store_type == 's3':
         from tilequeue.store import make_s3_store
@@ -572,7 +577,7 @@ def tilequeue_process(cfg, peripherals):
     assert n_simultaneous_query_sets > 0
     # reduce queue size when we're rendering metatiles to try and avoid the
     # geometry waiting to be processed from taking up all the RAM!
-    default_queue_buffer_size = max(1, 128 >> (2 * cfg.metatile_size))
+    default_queue_buffer_size = max(1, 128 >> (2 * (cfg.metatile_size or 0)))
     sql_queue_buffer_size = cfg.sql_queue_buffer_size or \
         default_queue_buffer_size
     proc_queue_buffer_size = cfg.proc_queue_buffer_size or \
@@ -628,8 +633,8 @@ def tilequeue_process(cfg, peripherals):
         post_process_data, formats, sql_data_fetch_queue, processor_queue,
         cfg.buffer_cfg, logger)
 
-    s3_storage = S3Storage(processor_queue, s3_store_queue, io_pool,
-                           store, logger, cfg.metatile_size, cfg.store_orig)
+    s3_storage = S3Storage(processor_queue, s3_store_queue, io_pool, store,
+                           logger, cfg.metatile_size)
 
     thread_sqs_writer_stop = threading.Event()
     sqs_queue_writer = SqsQueueWriter(sqs_queue, s3_store_queue, logger,
@@ -912,7 +917,11 @@ def tilequeue_seed(cfg, peripherals):
     if cfg.seed_should_add_to_tiles_of_interest:
         logger.info('Adding to Tiles of Interest ... ')
 
-        toi_set = peripherals.toi.fetch_tiles_of_interest()
+        if (cfg.toi_store_type == 'file' and
+                not os.path.exists(cfg.toi_store_file_name)):
+            toi_set = set()
+        else:
+            toi_set = peripherals.toi.fetch_tiles_of_interest()
 
         tile_generator = make_seed_tile_generator(cfg)
         for coord in tile_generator:
@@ -1118,12 +1127,34 @@ def tilequeue_prune_tiles_of_interest(cfg, peripherals):
                     coord_marshall_int(deserialize_coord(l.strip()))
                     for l in f
                 )
+        elif 'bucket' in info:
+            from boto import connect_s3
+            from boto.s3.bucket import Bucket
+            s3_conn = connect_s3()
+            bucket = Bucket(s3_conn, info['bucket'])
+            key = bucket.get_key(info['key'])
+            raw_coord_data = key.get_contents_as_string()
+            for line in raw_coord_data.splitlines():
+                coord = deserialize_coord(line.strip())
+                if coord:
+                    # NOTE: the tiles in the file should be of the
+                    # same size as the toi
+                    coord_int = coord_marshall_int(coord)
+                    immortal_tiles.add(coord_int)
 
         # Filter out nulls that might sneak in for various reasons
         immortal_tiles = filter(None, immortal_tiles)
 
         n_inc = len(immortal_tiles)
         new_toi = new_toi.union(immortal_tiles)
+
+        # ensure that the new coordinates have valid zooms
+        new_toi_valid_range = set()
+        for coord_int in new_toi:
+            coord = coord_unmarshall_int(coord_int)
+            if coord_is_valid(coord, cfg.max_zoom):
+                new_toi_valid_range.add(coord_int)
+        new_toi = new_toi_valid_range
 
         logger.info('Adding in tiles from %s ... done. %s found', name, n_inc)
 
@@ -1460,45 +1491,6 @@ def tilequeue_dump_tiles_of_interest(cfg, peripherals):
     )
 
 
-def tilequeue_dump_tiles_of_interest_from_redis(cfg, peripherals):
-    """
-    Dumps the tiles of interest from Redis into a newline-delimited
-    file called 'toi.txt'.
-
-    This is intended to be used once to migrate away from using Redis
-    and then removed.
-    """
-    logger = make_logger(cfg, 'dump_tiles_of_interest_from_redis')
-    logger.info('Dumping TOI from Redis ...')
-
-    logger.info('Fetching tiles of interest from Redis ...')
-
-    # Make a raw Redis client so we can do low-level set manipulation
-    redis_client = make_redis_client(cfg)
-
-    toi_iter = redis_client.sscan_iter(cfg.redis_cache_set_key, count=10000)
-
-    toi_set = set()
-    for coord_int in toi_iter:
-        toi_set.add(coord_int)
-
-    logger.info('Fetching tiles of interest from Redis ... done')
-
-    toi_filename = "toi.txt"
-
-    n_toi = len(toi_set)
-    logger.info('Writing %d tiles of interest to %s ...', n_toi, toi_filename)
-
-    with open(toi_filename, "w") as f:
-        save_set_to_fp(toi_set, f)
-
-    logger.info(
-        'Writing %d tiles of interest to %s ... done',
-        n_toi,
-        toi_filename
-    )
-
-
 def tilequeue_load_tiles_of_interest(cfg, peripherals):
     """
     Given a newline-delimited file containing tile coordinates in
@@ -1578,9 +1570,6 @@ def tilequeue_main(argv_args=None):
         ('intersect', create_command_parser(tilequeue_intersect)),
         ('dump-tiles-of-interest',
             create_command_parser(tilequeue_dump_tiles_of_interest)),
-        ('dump-tiles-of-interest-from-redis',
-            create_command_parser(
-                tilequeue_dump_tiles_of_interest_from_redis)),
         ('load-tiles-of-interest',
             create_command_parser(tilequeue_load_tiles_of_interest)),
         ('enqueue-tiles-of-interest',